+++ changelog.md
@@ -3,6 +3,7 @@
   * Add mixins.amplify.
   * Add Lateralus.prototype.dispose.
   * Lateralus.Component#dispose triggers "beforeDispose" event.
+  * Extend Lateralus.Component.Model from Backbone.Model, not Lateralus.Model.
 
 # 0.5.0